+++ src/storage/index.ts
@@ -1,5 +1,6 @@
 export * from "./adapter.js";
 export * from "./repository.js";
+export * from "./sqlite/connection.js";
 export * from "./sqlite/migrations.js";
 export * from "./sqlite/sqlite-repository.js";
 export * from "./sqlite/schema.js";